Simulate a population and a study of it

Simulate a population and a study of it

Usage

sim_pop_study(
  phi,
  lambda,
  N.init,
  hist.len,
  srvy.yrs,
  k,
  f.year,
  p,
  L,
  imaf,
  clvng.p,
  tmp.emgn,
  alpha,
  clvng.ints
)

Arguments

phi

The individual survival rate

lambda

The population growth rate

N.init

Initial population size

hist.len

Length of simulation

srvy.yrs

Survey years

k

Number of survey years

f.year

Final year of simulation

p

Capture probability

L

Number of genotype loci

imaf

Initial minor allele frequency

clvng.p

Additional calving probability

tmp.emgn

Rate of temporary immigration

alpha

Age of sexual maturity

clvng.ints

Whether females should be selected for calving by time since last calving, producing intervals of standard length

Value

A dataframe with ID, mum, dad, capture, and calving columns, and several attributes.

Examples

# Simulate one population and study
pop_study = sim_pop_study(
  phi = 0.9, lambda = 1.05, N.init = 20, hist.len = 20, srvy.yrs = 20, k = 1,
  f.year = 20, p = 0.5, L = 10, imaf = 0.5, clvng.p = 0, tmp.emgn = 0,
  alpha = 5, clvng.ints = FALSE
)

# Look at it
head(pop_study)
#>    ID mum dad C20 Cvg20
#> 18 18  NA  NA   1     0
#> 21 21  12   2   1     1
#> 22 22   4   3   1     0
#> 32 32  19  17   1     0
#> 35 35  19  15   1     0
#> 46 46  22  10   1     0
names(attributes(pop_study))
#>  [1] "names"       "row.names"   "class"       "avg.phi.obs" "beta"       
#>  [6] "N.t.vec"     "ns.caps"     "Ns"          "ns.clvng"    "alive"      
#> [11] "alv.s.yrs"   "f.age"       "mum"         "dad"         "ID"         
#> [16] "ind.gts"